Badge LHEO - Remplissage des parcours - Méthodes pour le XML dans un service
